3214 - Merge Panel and EditForm
