+++ src/client/components/SlidingPanel/SlidingPanel.tsx
@@ -6,23 +6,22 @@ import SlidingPanelComponent from 'react-sliding-side-panel'
 import Icon from 'client/components/Icon'
 
 type Props = {
-  openPanel: boolean
   closePanel: () => void
+  opened: boolean
+  size?: number
 }
 
 const SlidingPanel: React.FC<React.PropsWithChildren<Props>> = (props) => {
-  const { children, openPanel, closePanel } = props
+  const { children, closePanel, opened, size } = props
   const panelType = 'right'
-  const panelSize = 30
 
   return (
     <SlidingPanelComponent
-      type={panelType}
-      isOpen={openPanel}
       backdropClicked={closePanel}
-      size={panelSize}
+      isOpen={opened}
       panelClassName="sliding-panel"
-      panelContainerClassName=""
+      size={size}
+      type={panelType}
     >
       <>
         <button className="btn-s btn btn-transparent sliding-panel__close" onClick={closePanel} type="button">
@@ -32,6 +31,10 @@ const SlidingPanel: React.FC<React.PropsWithChildren<Props>> = (props) => {
       </>
     </SlidingPanelComponent>
   )
+}
+
+SlidingPanel.defaultProps = {
+  size: 30,
 }
 
 export default SlidingPanel
