--- src/client/pages/CountryHome/Repository/Panel/Panel.scss
@@ -1,8 +1,0 @@
-@import 'src/client/style/partials';
-
-.repository-form__container {
-  display: grid;
-  grid-gap: $spacing-s;
-  grid-template-columns: 1fr;
-  padding: $spacing-m;
-}
