--- src/client/pages/CountryHome/Repository/Panel/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Panel'
